refactor out the parameter subscription part such that we can re-use it on the sender and receiver side
