+++ src/mavsdk/core/CMakeLists.txt
@@ -25,6 +25,7 @@ target_sources(mavsdk
     mavlink_mission_transfer.cpp
     mavlink_parameter_receiver.cpp
     mavlink_parameter_sender.cpp
+    mavlink_parameter_subscription.cpp
     mavlink_receiver.cpp
     mavlink_request_message_handler.cpp
     mavlink_statustext_handler.cpp
